Update CustomHashMap.java
Updated interface to turn hashfunctions into long values to avoid overflow errors

+++ CustomHashMap.java
@@ -3,8 +3,8 @@ package group_project;
 public interface CustomHashMap<T, U> {
 
     void resize(int size);
-    int hashfunction(U key);
-    int hash2(int key);
+    long hashfunction(U key);
+    long hash2(long key);
     void put(T value, U key);
     GenKeyVal<T, U> get(U key);
